Created src folder for textcraft functions. Fixed some logic to match the desired outputs of pre-defined tests
