+++ src/textcraft/__init__.py
@@ -1,0 +1,4 @@
+from .textcraft import *
+
+__all__ = ["to_lowercase", "to_uppercase", "to_snake_case", "to_camel_case", "to_kebab_case",
+           "remove_punctuation", "normalize_spaces", "word_count", "char_count", "sentence_count", "slugify"]
